+++ test/test_Physics_Model.jl
@@ -191,6 +191,13 @@
 						
 		end #testset	
 		
+	@testset "source" 	begin
+		SetSrcToPoint(true)
+		pnt1 = Point(5)
+		@test source(pnt1) == Tuple(pnt1, 0.0, 0.0)
+	
+	end
+		
 	@testset "Invalid Detector Dimention $dim"  for dim =  
 	    Number[0, -1, 0//1, -1//1, -e, 0.0, -1.0, -Inf, Inf,]